move jackson mapper to own class
